+++ src/compas/scene/sceneobject.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 
 from abc import abstractmethod
 from .descriptors.protocol import DescriptorProtocol
+from .descriptors.colordict import ColorDictAttribute
 from .context import clear
 
 
@@ -26,10 +27,23 @@ class SceneObject(object):
     # add this to support the descriptor protocol vor Python versions below 3.6
     __metaclass__ = DescriptorProtocol
 
+    color = ColorDictAttribute()
+    pointcolor = ColorDictAttribute()
+    linecolor = ColorDictAttribute()
+    facecolor = ColorDictAttribute()
+
     def __init__(self, item, **kwargs):
         self._item = item
         self._transformation = None
         self._guids = None
+
+        self.color = kwargs.get("opacity", self.color)
+        self.pointcolor = kwargs.get("pointcolor", self.pointcolor)
+        self.linecolor = kwargs.get("linecolor", self.linecolor)
+        self.facecolor = kwargs.get("facecolor", self.facecolor)
+        self.pointsize = kwargs.get("pointsize", 1.0)
+        self.linewidth = kwargs.get("linewidth", 1.0)
+        self.opacity = kwargs.get("opacity", 1.0)
 
     @property
     def guids(self):